feat: make availability page responsive
