+++ src/modules/event/components/EventItem.tsx
@@ -1,43 +1,17 @@
 import Link from 'next/link';
 import { toast } from 'sonner';
 import { useTransition } from 'react';
-import {
-  LinkIcon,
-  ClockIcon,
-  TrashIcon,
-  ExternalLinkIcon,
-  EllipsisVerticalIcon,
-} from 'lucide-react';
+import { LinkIcon, ClockIcon, ExternalLinkIcon } from 'lucide-react';
 
 import { Button } from '@/shared/components/ui/button';
 import { Skeleton } from '@/shared/components/ui/skeleton';
 import { Heading } from '@/shared/components/layout/Heading';
 import { useUser } from '@/modules/auth/contexts/UserContext';
-import { StatefulButton } from '@/shared/components/ui/stateful-button';
 import {
   Tooltip,
   TooltipContent,
   TooltipTrigger,
 } from '@/shared/components/ui/tooltip';
-import {
-  DropdownMenu,
-  DropdownMenuItem,
-  DropdownMenuContent,
-  DropdownMenuTrigger,
-} from '@/shared/components/ui/dropdown-menu';
-import {
-  AlertDialog,
-  AlertDialogTitle,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogContent,
-  AlertDialogTrigger,
-  AlertDialogDescription,
-} from '@/shared/components/ui/alert-dialog';
-
-import { useEventMutations } from '../hooks/useEventMutations';
 
 type EventItemProps = {
   event: {
@@ -78,15 +52,9 @@ type EventItemActionsProps = {
 
 function EventItemActions({ event }: EventItemActionsProps) {
   const { slug: hostSlug } = useUser();
-  const { deleteEvent, isDeletingEvent } = useEventMutations();
   const [isCopyingLink, startCopyingLink] = useTransition();
 
   const eventPath = `/sharing/${hostSlug}/${event.slug}`;
-
-  async function handleDeleteEvent() {
-    if (isDeletingEvent) return;
-    await deleteEvent(event.id);
-  }
 
   async function handleCopyLink() {
     const eventLink = `${window.location.origin}${eventPath}`;
@@ -133,52 +101,6 @@ function EventItemActions({ event }: EventItemActionsProps) {
           <p>Copy link to event</p>
         </TooltipContent>
       </Tooltip>
-
-      <AlertDialog>
-        <DropdownMenu>
-          <DropdownMenuTrigger asChild>
-            <Button
-              size="icon"
-              variant="ghost"
-              className="bg-background rounded-none"
-              aria-label="Delete event"
-            >
-              <EllipsisVerticalIcon className="size-4" />
-            </Button>
-          </DropdownMenuTrigger>
-          <DropdownMenuContent align="end">
-            <AlertDialogTrigger asChild>
-              <DropdownMenuItem>
-                <TrashIcon />
-                Delete
-              </DropdownMenuItem>
-            </AlertDialogTrigger>
-          </DropdownMenuContent>
-        </DropdownMenu>
-        <AlertDialogContent>
-          <AlertDialogHeader>
-            <AlertDialogTitle>Delete Event</AlertDialogTitle>
-            <AlertDialogDescription>
-              Are you sure you want to delete this event?
-            </AlertDialogDescription>
-          </AlertDialogHeader>
-          <AlertDialogFooter>
-            <AlertDialogCancel asChild>
-              <Button variant="outline" disabled={isDeletingEvent}>
-                Cancel
-              </Button>
-            </AlertDialogCancel>
-            <AlertDialogAction
-              asChild
-              variant={'destructive'}
-              disabled={isDeletingEvent}
-              onClick={handleDeleteEvent}
-            >
-              <StatefulButton loading={isDeletingEvent}>Delete</StatefulButton>
-            </AlertDialogAction>
-          </AlertDialogFooter>
-        </AlertDialogContent>
-      </AlertDialog>
     </div>
   );
 }
@@ -194,7 +116,6 @@ export function EventItemFallback() {
         <Skeleton className="h-5 w-[8ch]" />
       </div>
       <div className="flex items-center gap-[1px] overflow-clip rounded-md">
-        <Skeleton className="size-9 rounded-none" />
         <Skeleton className="size-9 rounded-none" />
         <Skeleton className="size-9 rounded-none" />
       </div>
